Add Feedback tests

+++ src/com/practice/projectone/teammanagement/models/tasks/contracts/Feedback.java
@@ -1,6 +1,6 @@
 package com.practice.projectone.teammanagement.models.tasks.contracts;
 
 public interface Feedback extends Task{
-    double getRating();
+    int getRating();
     public void changeRating(int rating);
 }
